Removed experimental modules feature.
