+++ stroom-app/src/main/resources/ui/.gitignore
@@ -1,2 +1,3 @@
 /stroom/
 /dashboard/
+/swagger/
