--- stroom-core-module/.gitignore
@@ -1,7 +1,0 @@
-/.classpath
-/.externalToolBuilders/
-/.project
-/.settings/
-/target/
-/build/
-/test-output/
