--- stroom-core-module/build.gradle
@@ -1,74 +1,0 @@
-plugins {
-    id "com.benjaminsproule.swagger" version "0.1.4"
-}
-
-def warDir = project(':stroom-app').file('src/main/resources/ui')
-def outputDir = "${warDir}/swagger"
-
-dependencies {
-    compile project(':stroom-module')
-
-    compile project(':stroom-core-shared')
-    compile project(':stroom-core-server')
-    compile project(':stroom-security-server')
-    compile project(':stroom-index-server')
-    compile project(':stroom-statistics-api')
-    permitUnusedDeclared project(':stroom-statistics-api') // not clear why it thinks this is unused
-    compile project(':stroom-util')
-    compile project(':stroom-pipeline')
-    compile project(':stroom-dashboard-server')
-    compile project(':stroom-statistics-server')
-
-    compile urlDependencies.get(urlLibs.stroomQueryApi)
-    compile urlDependencies.get(urlLibs.stroomQueryCommon)
-
-    compile(libs.dropwizard_core) {
-        exclude(group: "org.glassfish.web", module: "javax.el")
-    }
-    compile libs.dropwizard_lifecycle
-    compile libs.dropwizard_jersey
-    compile libs.dropwizard_jetty
-    compile libs.dropwizard_metrics_healthchecks
-    compile libs.guava
-    compile libs.javax_inject
-    compile libs.javax_servlet_api
-    compile libs.jetty_servlet
-    compile libs.shiro_web
-    compile libs.slf4j_api
-    compile libs.spring_core
-    compile libs.spring_beans
-    compile libs.spring_context
-}
-
-swagger {
-    apiSource {
-        springmvc = false //makes it search for jersey annotations
-        outputFormats = 'json,yaml'
-        locations = [
-                'stroom.statistics.server.sql.search.SqlStatisticsQueryResource',
-                'stroom.index.server.StroomIndexQueryResource',
-                'stroom.security.server.AuthorisationResource' ]
-        schemes = [
-                'http',
-                'https' ]
-        host = 'localhost:8080'
-        basePath = '/api'
-        info {
-            title = 'stroom API'
-            version = 'v1'
-            description = 'Various APIs for interacting with stroom and its data'
-            contact {
-                name = 'GCHQ'
-                url = 'https://github.com/gchq/stroom'
-            }
-            license {
-                url = 'http://www.apache.org/licenses/LICENSE-2.0.html'
-                name = 'Apache 2.0'
-            }
-        }
-        outputPath = "${project.buildDir}/generated/document.html"
-        swaggerDirectory = "${outputDir}"
-    }
-}
-
-tasks.build.dependsOn generateSwaggerDocumentation
