--- stroom-core-module/src/main/resources/META-INF/services/stroom.module.StroomModule
@@ -1,1 +1,0 @@
-stroom.module.StroomCoreModule    # Core Application Module
