--- stroom-module/build.gradle
@@ -1,5 +1,0 @@
-dependencies {
-    compile (libs.dropwizard_core) {
-        exclude ( group: "org.glassfish.web", module: "javax.el" )
-    }
-}
