--- stroom-module/src/main/java/stroom/module/StroomModule.java
@@ -1,8 +1,0 @@
-package stroom.module;
-
-import io.dropwizard.Configuration;
-import io.dropwizard.setup.Environment;
-
-public interface StroomModule {
-    void initialize(Configuration configuration, Environment environment);
-}
